Moved wrapper classes to internal package
