+++ src/main/java/com/vonage/client/messages/MessageRequest.java
@@ -20,6 +20,9 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vonage.client.VonageUnexpectedException;
+import com.vonage.client.messages.internal.Channel;
+import com.vonage.client.messages.internal.E164;
+import com.vonage.client.messages.internal.MessageType;
 
 import java.util.Objects;
