+++ src/main/java/com/vonage/client/messages/internal/MessagePayload.java
@@ -13,7 +13,7 @@
  *   See the License for the specific language governing permissions and
  *   limitations under the License.
  */
-package com.vonage.client.messages;
+package com.vonage.client.messages.internal;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
