+++ src/main/java/com/vonage/client/messages/messenger/MessengerRequest.java
@@ -16,7 +16,7 @@
 package com.vonage.client.messages.messenger;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.vonage.client.messages.Channel;
+import com.vonage.client.messages.internal.Channel;
 import com.vonage.client.messages.MessageRequest;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
