+++ src/main/java/com/vonage/client/messages/sms/SmsRequest.java
@@ -17,10 +17,10 @@ package com.vonage.client.messages.sms;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.vonage.client.messages.Channel;
+import com.vonage.client.messages.internal.Channel;
 import com.vonage.client.messages.MessageRequest;
-import com.vonage.client.messages.MessageType;
-import com.vonage.client.messages.Text;
+import com.vonage.client.messages.internal.MessageType;
+import com.vonage.client.messages.internal.Text;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 public final class SmsRequest extends MessageRequest {
