+++ src/test/java/com/vonage/client/messages/MessageRequestTest.java
@@ -18,6 +18,8 @@ package com.vonage.client.messages;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vonage.client.VonageUnexpectedException;
+import com.vonage.client.messages.internal.Channel;
+import com.vonage.client.messages.internal.MessageType;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
